# PERF: Specialize loadtxt packer for uniform-dtype data.