~15-30% speedup.  (I'm making these one at a time to be able to show the performance gain of each of them separately.)

This specializes _loadtxt_pack_items (a generate nested tuple/list packer that handles arbitrarily nested dtypes) for the common case where there's just a scalar dtype, in which case the packer is just a no-op.

I finally understood how to use --bench-compare :-)
```
       before           after         ratio
     [b5cdbf04]       [78e39da0]
     <main>           <loadtxtspecialpacker>
-      53.6±0.2μs       44.9±0.2μs     0.84  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('complex128', 10)
-      49.3±0.4μs       39.6±0.4μs     0.80  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('str', 10)
-      45.6±0.5μs       36.5±0.2μs     0.80  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('object', 10)
-      46.3±0.3μs       37.0±0.1μs     0.80  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('float64', 10)
-      47.1±0.1μs       37.5±0.2μs     0.80  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('int32', 10)
-      46.5±0.3μs      37.0±0.05μs     0.80  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('float32', 10)
-         365±2ms          289±2ms     0.79  bench_io.LoadtxtCSVSkipRows.time_skiprows_csv(500)
-         331±3ms          263±1ms     0.79  bench_io.LoadtxtCSVSkipRows.time_skiprows_csv(10000)
-      45.9±0.3μs       36.3±0.2μs     0.79  bench_io.LoadtxtCSVComments.time_comment_loadtxt_csv(10)
-         366±2ms          289±2ms     0.79  bench_io.LoadtxtCSVSkipRows.time_skiprows_csv(0)
-         359±2μs          283±3μs     0.79  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('complex128', 100)
-      34.3±0.1ms       26.9±0.5ms     0.78  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('complex128', 10000)
-       354±0.9ms          277±3ms     0.78  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('complex128', 100000)
-      48.5±0.5μs       38.0±0.4μs     0.78  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('int64', 10)
-      16.8±0.1ms      13.1±0.05ms     0.78  bench_io.LoadtxtUseColsCSV.time_loadtxt_usecols_csv([1, 3, 5, 7])
-         296±1μs        221±0.8μs     0.75  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('int32', 100)
-     14.2±0.07ms       10.5±0.1ms     0.74  bench_io.LoadtxtUseColsCSV.time_loadtxt_usecols_csv([1, 3])
-         316±1μs          235±3μs     0.74  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('str', 100)
-         290±4μs        214±0.9μs     0.74  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('float64', 100)
-         309±1μs          227±1μs     0.74  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('int64', 100)
-         290±2μs        213±0.8μs     0.74  bench_io.LoadtxtCSVComments.time_comment_loadtxt_csv(100)
-         286±3μs          210±2μs     0.73  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('object', 100)
-      29.9±0.2ms       22.0±0.3ms     0.73  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('str', 10000)
-       300±0.9ms          220±2ms     0.73  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('str', 100000)
-       292±0.9μs        213±0.9μs     0.73  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('float32', 100)
-         288±3ms        210±0.6ms     0.73  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('int32', 100000)
-      27.0±0.1ms       19.6±0.1ms     0.73  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('object', 10000)
-      27.4±0.2ms       19.9±0.1ms     0.73  bench_io.LoadtxtCSVComments.time_comment_loadtxt_csv(10000)
-         279±2ms          202±2ms     0.72  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('object', 100000)
-      29.7±0.1ms       21.5±0.1ms     0.72  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('int64', 10000)
-         286±2ms          207±1ms     0.72  bench_io.LoadtxtCSVComments.time_comment_loadtxt_csv(100000)
-         285±3ms          206±1ms     0.72  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('float64', 100000)
-         308±2ms          223±1ms     0.72  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('int64', 100000)
-      27.4±0.4ms      19.7±0.07ms     0.72  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('float32', 10000)
-      27.4±0.2ms       19.7±0.2ms     0.72  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('float64', 10000)
-      28.5±0.5ms      20.4±0.06ms     0.72  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('int32', 10000)
-         288±3ms        206±0.7ms     0.72  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('float32', 100000)
```