@mattip 2021-08-04
Cool, thanks @anntzer 

At some point it would be nice to get a release note in `doc/release/upcoming_changes` with a short mention that loadtxt is now faster.