@anntzer 2021-08-04
I actually have a few more of these in, I'll do a big benchmark comparison at the end :-)